requirements.txt
# 核心依赖
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
pillow>=10.0.0

# 动态化学知识库（推荐）
pymatgen>=2024.0.0

# 可选：如果不想使用 pymatgen，可以使用 pubchempy
# pubchempy>=1.0.4
streamlit
tiktoken

# 外部文献 RAG 依赖（可选，用于增强专业性）
chromadb>=0.4.0
pdfplumber>=0.10.0
tqdm>=4.65.0

# sk-eec9cb28d6804d18aaddcdb4bdd9a1b9